Expose AEAD_SM4_GCM && AEAD_SM4_CCM

+++ src/blockmode/ccm.rs
@@ -35,9 +35,10 @@ macro_rules! impl_block_cipher_with_ccm_mode {
             pub const TAG_LEN: usize   = $tlen;
             
             #[cfg(target_pointer_width = "64")]
-            pub const A_MAX: usize = usize::MAX as usize; // 2^64 - 1
+            pub const A_MAX: usize = usize::MAX;                  // 2^64 - 1
             #[cfg(target_pointer_width = "32")]
-            pub const A_MAX: usize = usize::MAX;          // 2^32 - 1
+            pub const A_MAX: usize = usize::MAX;                  // 2^32 - 1
+            
             pub const P_MAX: usize = 16777215;                    // 2^24 - 1
             pub const C_MAX: usize = Self::P_MAX + Self::TAG_LEN; // 2^24 + 15
